size optimization by switching to single file components
using .vue files, vite is able to prebuild the templates, which removes the need for the runtime vue compiler: shrinking by about 100kB ( and about 77kB smaller than /calendar )
