+++ cal/src/App.vue
@@ -1,0 +1,18 @@
+<!-- 
+  boilerplate for using vue-router
+  if we refactored to share the "chrome" of the overview and detail pages that chrome would go here.
+ -->
+<script>
+import { RouterView } from 'vue-router'
+
+export default {
+  components:{ RouterView }
+}
+</script>
+
+<template>
+  <router-view></router-view>
+</template>
+
+<style>
+</style>
