+++ cal/src/DelSpan.vue
@@ -1,10 +1,9 @@
+<script>
 /*
  * if we need it, an example of swapping b/t <del> and <span>
  * ex. <DelSpan deleted=true>content</DelSpan> ==> `<del>content</del>
  */
 export default {
-// "slot" is the user content between the tags
-template: `<component :is="htmlEl"><slot></slot></component>`,
   computed: {
     htmlEl() {
       return this.deleted ? "del" : "span";
@@ -16,4 +15,13 @@ template: `<component :is="htmlEl"><slot></slot></component>`,
       required: true,
     }
   }
-};
+}
+</script>
+
+<!-- "slot" is the user content between the tags -->
+<template>
+  <component :is="htmlEl"><slot></slot></component>
+</template>
+
+<style>
+</style>
