+++ cal/src/QuickNav.vue
@@ -1,3 +1,4 @@
+<script>
 /**
  * QuickNav provides a row of buttons (at the bottom of a page)
  * for navigating through events, with common shortcuts 
@@ -9,20 +10,6 @@
  * shortcut: { id, icon, label, (emit|url) }
  */ 
 export default {
-  template: `
-<div class="c-shortcuts">
-  <div class="c-shortcut" v-for="el in shortcuts" :key="el.id">
-    <a v-if="el.url"  class="c-shortcut__link"
-      :href="el.url" 
-      v-bind="el.attrs"
-    >{{el.icon}}</a>
-    <button v-else class="c-shortcut__button" 
-      @click="onClick(el)"
-      v-bind="el.attrs"
-    >{{el.icon}}</button>
-    <div class="c-shortcut__label">{{el.label}}</div>
-  </div>
-</div>`,
   props: {
     shortcuts: {
       type: Array,
@@ -49,3 +36,23 @@ export default {
     }
   }
 }
+</script>
+
+<template>
+  <div class="c-shortcuts">
+  <div class="c-shortcut" v-for="el in shortcuts" :key="el.id">
+    <a v-if="el.url"  class="c-shortcut__link"
+      :href="el.url" 
+      v-bind="el.attrs"
+    >{{el.icon}}</a>
+    <button v-else class="c-shortcut__button" 
+      @click="onClick(el)"
+      v-bind="el.attrs"
+    >{{el.icon}}</button>
+    <div class="c-shortcut__label">{{el.label}}</div>
+  </div>
+  </div>
+</template>
+
+<style>
+</style>
