+++ cal/src/main.js
@@ -1,21 +1,14 @@
 /**
  * entry point for the single page application.
- * calMain.js gets loaded by '/events/index.html'
- * 
- * generated by hugo via 'site/content/events.md' and 'site/layouts/events/single.html'
- * we could drop hugo for events IF we can still generate menu and pedalp data from hugo
- * https://gohugo.io/content-management/content-adapters/
- * 
- * in production mapping individual event to the events page 
- * using 'npm run dev' that's handled by facade.js
  */
 import { createApp } from 'vue'
-import { createWebHistory, createRouter, RouterView } from 'vue-router'
+import { createWebHistory, createRouter } from 'vue-router'
 import './style.css'
 
 // our main pages:
-import CalPage from './calPage.js'
-import SingleEvent from './singleEvent.js'
+import App from './App.vue'
+import CalPage from './CalPage.vue'
+import EventDetails from './EventDetails.vue'
 
 // the router reads and writes the user's address bar
 const router = createRouter({
@@ -30,21 +23,17 @@ const router = createRouter({
     }, 
     // named elements are used by the javascript code to link from place to place
     { 
-      name: 'singleEvent', 
+      name: 'EventDetails', 
       // caldaily_id will only match numbers
       // slug is optional
       // tbd: any next and prev events might want to use ?rel=next, ?rel=prev
       // to avoid potential conflict with the slug
       path: '/events/:caldaily_id(\\d+)/:slug?', 
-      component: SingleEvent 
+      component: EventDetails 
     },
   ],
 })
 
-createApp({
-  components:{
-    RouterView,
-  },
-  template: ` <router-view></router-view>`})
+createApp(App)
 .use(router)
 .mount('#app');
